Fix CHW mixup with images
All of the SAM3 notebooks and examples use a PIL image, and this works,
but some portions of SAM3 assume HWC format and others assume CHW for
numpy arrays. Within the `set_image` method for the `Sam3Processor`, it
is even confused by this order because they are assuming CHW for the
image size determination but then later transpose it to CHW, which flips
everything around when moving to a tensor and causes an error.

Like other parts of the code, we will assume the numpy arrays are in
HWC, which is from OpenCV-python numpy arrays and makes sense. Tensors
are a different matter.